census: which rail a line describes is the running rail, not the compiled one
`drain/census` chose its rail lines with `cfg`, and one of those gates was
already wrong for the goal of this branch: the Metal cache-levels emitter was
spelled `all(not(backend-vulkan), backend-metal, macos)`. Turning both features
on does not make that line ambiguous — it makes it **vanish**, silently, and
the only symptom would be one census line missing from a log nobody was
diffing.

The others fail the other way. `emit_engine_delta`, `emit_engine_lock`, the two
working-set lines and `emit_guest_import_levels` all read
`backend::vulkan::engine`, so on a both-rails build running Metal they would
report the Vulkan engine's counters — and take its mutex — for a device that
never touched it. `cfg` answers "which rail was compiled"; the question here
has always been "which rail is running".

`Backend::emit_census(CensusSite)` asks the running rail. `CensusSite` names the
four points in the drain's window at which a rail may speak, and it is named for
the *drain's* reason at each — `Serialization { win_ms }` under
`window_publish`, `WorkingSet` beside the eviction routes, `Throughput` before
`chain_phase`, `Levels` after it — because several neutral lines between them
are only interpretable read against a rail line that must come first. Ordering
between sites is the drain's; ordering within one is the rail's, which is why
`engine_delta` → `registry_pressure` → `draw_phase` stays intact inside the
rail.

`census.rs` becomes `census/{mod,vulkan,metal}.rs`, matching `draw/`. The bodies
move unchanged. A rail with nothing to say at a site says nothing, which is not
a zero: an absent `engine_delta` means no such engine where `engine_delta …=0`
would mean an idle one.

The drain resolves the rail once per window, so four asks cannot land on two
rails.

Behaviour on each single-rail build is unchanged, line for line and in the same
order. Verified on both arms: feature matrix, clippy `-D warnings`, full suites
serially (metal 1326, vulkan 2269, support 351, unchanged).
